requirements.txt
channels[daphne]
channels_redis
dj-database-url
django
djangorestframework
pip
psycopg[c]
psycopg_pool
